requirements.txt
gunicorn==20.1.0
django==3.2.19
djangorestframework==3.14.0
pillow==9.5.0
django-heroku==0.3.1
django-cors-headers==4.0.0
django-debug-toolbar
psycopg2-binary==2.9.6
channels==4.0.0
channels-redis==4.1.0
django-redis==5.2.0
django-environ==0.10.0
django-filter==23.2
django-dynamic-fixture==3.1.2
daphne==4.0.0
celery==5.2.7
importlib-metadata==4.12

